dummy-backup/app/requirements.txt
absl-py==1.4.0
asttokens==2.0.8
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
Automat==20.2.0
autopep8==2.0.0
backcall==0.2.0
backports.entry-points-selectable==1.1.1
beautifulsoup4==4.10.0
cachetools==5.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.6
click==8.0.2
cloudpickle==1.6.0
colorama==0.4.4
constantly==15.1.0
cryptography==36.0.0
cssselect==1.1.0
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
distlib==0.3.3
dnspython==2.2.1
dominate==2.7.0
email-validator==1.3.0
entrypoints==0.4
executing==1.0.0
filelock==3.4.0
flappy-bird-gym==0.3.0
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
flask-ngrok==0.0.25
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
flatbuffers==23.3.3
fonttools==4.37.1
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.1
grpcio==1.53.0
gTTS==2.2.3
gym==0.18.3
h11==0.12.0
h2==3.2.0
h5py==3.8.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.2
importlib-metadata==4.12.0
incremental==21.3.0
ipykernel==6.15.2
ipython==8.4.0
itemadapter==0.4.0
itemloaders==1.0.4
itsdangerous==2.1.2
jax==0.4.8
jedi==0.18.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.2.0
jupyter-core==4.11.1
jupyter_client==7.3.5
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
lxml==4.6.4
Markdown==3.4.3
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
ml-dtypes==0.1.0
myenv==201902.7
mysql-connector==2.2.9
nest-asyncio==1.5.5
nltk==3.7
numpy==1.23.5
oauthlib==3.2.0
opt-einsum==3.3.0
outcome==1.1.0
packaging==21.3
pandas==1.4.4
parsel==1.6.0
parso==0.8.3
pickleshare==0.7.5
Pillow==8.2.0
platformdirs==2.4.0
priority==1.3.0
prompt-toolkit==3.0.30
Protego==0.1.16
protobuf==4.22.3
psutil==5.9.1
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyChalk @ git+https://github.com/Throvn/pychalk@b78493cfc1797774b09ea0b4632f83e558efc448
pycodestyle==2.9.1
pycparser==2.21
PyDispatcher==2.0.5
pygame==2.0.3
pyglet==1.5.15
Pygments==2.13.0
PyMySQL==1.0.2
pyOpenSSL==21.0.0
pyparsing==3.0.9
pypng==0.0.21
PyQRCode==1.2.1
python-dateutil==2.8.2
pytz==2022.2.1
pywin32==304
pyzmq==23.2.1
queuelib==1.6.2
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.1
scipy==1.9.1
Scrapy==2.5.1
seaborn==0.11.2
selenium==4.1.0
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SQLAlchemy==1.4.44
stack-data==0.5.0
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.1.0
textblob==0.17.1
threadpoolctl==3.1.0
tk==0.1.0
tomli==2.0.1
tornado==6.2
tqdm==4.64.1
traitlets==5.3.0
trio==0.19.0
trio-websocket==0.9.2
tweepy==4.10.1
Twisted==21.7.0
twisted-iocpsupport==1.0.2
typing_extensions==4.0.1
urllib3==1.26.7
virtualenv==20.10.0
visitor==0.1.3
w3lib==1.22.0
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.0.0
WTForms==3.0.1
zipp==3.8.1
zope.interface==5.4.0
